chore: 폴더 변경 및 명칭 수정, 프로퍼티 적용방식 변경
- 각 모델 요청별 구분
- UploadArticle
- body 프로퍼티 추가
