+++ Targets/API/Sources/Response/Article/UpdateArticleResponse.swift
@@ -1,6 +1,6 @@
 //
-//  UploadArticleResponse.swift
-//  Ticlemoa
+//  UpdateArticleResponse.swift
+//  API
 //
 //  Created by 김우성 on 2022/12/22.
 //  Copyright © 2022 nyongnyong. All rights reserved.
@@ -8,7 +8,7 @@
 
 import Foundation
 
-public struct UploadArticleResponse: Decodable {
+public struct UpdateArticleResponse: Decodable {
     let content: String
     let isPublic: Bool
     let title: String
