+++ Targets/API/Sources/Response/Login/KakaoLoginResponse.swift
@@ -1,6 +1,6 @@
 //
 //  KakaoLoginResponse.swift
-//  Ticlemoa
+//  API
 //
 //  Created by 김우성 on 2022/12/21.
 //  Copyright © 2022 nyongnyong. All rights reserved.
@@ -9,10 +9,12 @@
 import Foundation
 
 public struct KakaoLoginResponse: Decodable {
-    public let accessToken: String
+    let accessToken: String
+    let userId: Int
     
-    init(accessToken: String) {
+    init(accessToken: String, userId: Int) {
         self.accessToken = accessToken
+        self.userId = userId
     }
     
 }
